requirements.txt
aniso8601==6.0.0
APScheduler==3.6.3
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==2.0.3
django-admin-json-editor==0.2.3
django-appconf==1.0.4
django-choices==1.7.1
django-colorfield==0.4.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-graphql-jwt==0.3.1
django-graphql-social-auth==0.1.4
django-image-cropping==1.2.0
django-money==1.3
django-phonenumber-field==3.0.1
django-summernote==0.8.11.6
djangorestframework==3.9.4
easy-thumbnails==2.7
graphene==2.1.6
graphene-django==2.3.2
graphene-file-upload==1.2.2
graphql-core==2.2
graphql-relay==0.4.5
gunicorn==19.6.0
idna==2.10
oauthlib==3.1.0
phonenumbers==8.10.14
Pillow==6.1.0
promise==2.3
psycopg2==2.8.4
py-moneyed==0.8.0
pycparser==2.20
pycrypto==2.6.1
python-telegram-bot==13.2
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
Rx==3.1.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
PyJWT==1.7.1
